# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KNXVault API
  version: 0.4.7
  description: Secrets management and PKI REST API (Phase 1 MVP + Phase 2 complete).
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8200
    description: Local development
...
paths:
...
  /sys/policies:
    get:
      tags: [sys]
      summary: List RBAC policies
      operationId: listPolicies
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Policy list }
  /sys/policies/{name}:
    put:
      tags: [sys]
      summary: Create or update RBAC policy
      operationId: putPolicy
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyRequest"
      responses:
        "204": { description: Policy saved }
    get:
      tags: [sys]
      summary: Get RBAC policy
      operationId: getPolicy
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
    delete:
      tags: [sys]
      summary: Delete RBAC policy
      operationId: deletePolicy
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Policy deleted }
  /sys/roles:
    get:
      tags: [sys]
      summary: List role bindings
      operationId: listRoles
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleResponse"
  /sys/roles/{name}:
    put:
      tags: [sys]
      summary: Create or update role binding
      operationId: putRole
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleRequest"
      responses:
        "204": { description: Role saved }
    get:
      tags: [sys]
      summary: Get role binding
      operationId: getRole
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleResponse"
    delete:
      tags: [sys]
      summary: Delete role binding
      operationId: deleteRole
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Role deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    PolicyRequest:
      type: object
      required: [effect, resources, actions]
      properties:
        effect: { type: string }
        resources:
          type: array
          items: { type: string }
        actions:
          type: array
          items: { type: string }
        conditions:
          type: object
          additionalProperties: true
    PolicyResponse:
      allOf:
        - $ref: "#/components/schemas/PolicyRequest"
        - type: object
          properties:
            name: { type: string }
    RoleRequest:
      type: object
      required: [policies]
      properties:
        policies:
          type: array
          items: { type: string }
    RoleResponse:
      type: object
      properties:
        name: { type: string }
        policies:
          type: array
          items: { type: string }
...